добавляет отдельный модуль для функций состояния отправки формы
