+++ js/main.js
@@ -5,7 +5,7 @@ import './validationForm.js';
 import './scale.js';
 import './effects.js';
 import {getData} from './api.js';
-
+import './conditionSending.js';
 
 getData((photos) => {
   renderPhotos(photos);
